Fix Swiglu test

+++ onnxruntime/contrib_ops/cuda/llm/cutlass_extensions/gemm/kernel/moe_cutlass_kernel.h
@@ -359,7 +359,7 @@ struct MoeFCGemm {
       return Status::kInvalid;
     }
     // Handle the case the input is too short
-    else if (args.gemm_n < Mma::IteratorB::AccessType::kElements) {
+    else if (args.gemm_n < static_cast<int64_t>(Mma::IteratorB::AccessType::kElements)) {
       CUTLASS_TRACE_HOST("MoeFCGemm::can_implement() - gemm_n is smaller than the input alignment");
       return Status::kInvalid;
     }
